Handle missing tests.
This avoids hanging when there are tests in the testlist that are
not implemented by the given deqp executable.

deqp skips all the tests it does not know about and hence will
eventually start exiting after executing 0 tests. This code
detects that and marks all remaining tests as missing.

Multiple people hits this as a mystery hang, so this should be
a significant usability improvement.

Reviewed-by: Samuel Pitoiset <samuel.pitoiset@gmail.com>